+++ R/rank_ANOVA.R
@@ -1,20 +1,19 @@
 #' Effect Size for Rank Based ANOVA
 #'
-#' Compute rank epsilon squared (\eqn{\varepsilon^2}{\epsilon^2}) (to accompany
-#' [stats::kruskal.test()]), and Kendall's *W* (to accompany
-#' [stats::friedman.test()]) effect sizes for non-parametric (rank sum) one-way
-#' ANOVAs.
+#' Compute rank epsilon squared (\eqn{E^2_R}) or rank eta squared
+#' (\eqn{\eta^2_H}) (to accompany [stats::kruskal.test()]), and Kendall's *W*
+#' (to accompany [stats::friedman.test()]) effect sizes for non-parametric (rank
+#' sum) one-way ANOVAs.
 #'
-#' @inheritParams phi
 #' @inheritParams rank_biserial
 #' @param x Can be one of:
 #'   - A numeric vector, or a character name of one in `data`.
-#'   - A list of vectors (for `rank_epsilon_squared()`).
+#'   - A list of vectors (for `rank_eta/epsilon_squared()`).
 #'   - A matrix of `blocks x groups` (for `kendalls_w()`) (or `groups x blocks`
 #'   if `blocks_on_rows = FALSE`). See details for the `blocks` and `groups`
 #'   terminology used here.
 #'   - A formula in the form of:
-#'       - `DV ~ groups` for `rank_epsilon_squared()`.
+#'       - `DV ~ groups` for `rank_eta/epsilon_squared()`.
 #'       - `DV ~ groups | blocks` for `kendalls_w()` (See details for the
 #'       `blocks` and `groups` terminology used here).
 #' @param groups,blocks A factor vector giving the group / block for the
@@ -26,10 +25,10 @@
 #'
 #'
 #' @details
-#' The rank epsilon squared is appropriate for non-parametric tests of
-#' differences between 2 or more samples (a rank based ANOVA). See
-#' [stats::kruskal.test]. Values range from 0 to 1, with larger values
-#' indicating larger differences between groups.
+#' The rank epsilon squared and rank eta squared are appropriate for
+#' non-parametric tests of differences between 2 or more samples (a rank based
+#' ANOVA). See [stats::kruskal.test]. Values range from 0 to 1, with larger
+#' values indicating larger differences between groups.
 #' \cr\cr
 #' Kendall's *W* is appropriate for non-parametric tests of differences between
 #' 2 or more dependent samples (a rank based rmANOVA), where each `group` (e.g.,
@@ -40,7 +39,7 @@
 #' / higher agreement between raters.
 #'
 #' # Confidence (Compatibility) Intervals (CIs)
-#' Confidence intervals for rank Epsilon squared, and Kendall's *W* are
+#' Confidence intervals for \eqn{E^2_R}, \eqn{\eta^2_H}, and Kendall's *W* are
 #' estimated using the bootstrap method (using the `{boot}` package).
 #'
 #' @inheritSection effectsize_CIs CIs and Significance Tests
@@ -48,16 +47,17 @@
 #' @inheritSection rank_biserial Ties
 #'
 #'
-#' @return A data frame with the effect size `(`rank_epsilon_squared` or
-#'   `Kendalls_W`) and its CI (`CI_low` and `CI_high`).
+#' @return A data frame with the effect size and its CI.
 #'
 #' @family effect size indices
 #' @seealso [rank_biserial()] for more rank based effect sizes
 #'
 #' @examples
 #' \donttest{
-#' # Rank Epsilon Squared
-#' # ====================
+#' # Rank Eta/Epsilon Squared
+#' # ========================
+#'
+#' rank_eta_squared(mpg ~ cyl, data = mtcars)
 #'
 #' rank_epsilon_squared(mpg ~ cyl, data = mtcars)
 #'
@@ -76,6 +76,12 @@
 #' interpret(W, rules = "landis1977")
 #' }
 #'
+#' @references
+#' - Kendall, M.G. (1948) Rank correlation methods. London: Griffin.
+#'
+#' - Tomczak, M., & Tomczak, E. (2014). The need to report effect size estimates
+#' revisited. An overview of some recommended measures of effect size. Trends in
+#' sport sciences, 1(21), 19-25.
 #'
 #' @export
 #' @importFrom stats na.omit
@@ -93,7 +99,7 @@ rank_epsilon_squared <- function(x,
     if (!grepl("Kruskal-Wallis", x$method)) {
       stop("'x' is not a Kruskal-Wallis-test!", call. = FALSE)
     }
-    return(effectsize(x, ci = ci, iterations = iterations, alternative = alternative))
+    return(effectsize(x, type = "epsilon", ci = ci, iterations = iterations, alternative = alternative))
   }
 
   ## pep data
@@ -104,16 +110,20 @@ rank_epsilon_squared <- function(x,
   out <- data.frame(rank_epsilon_squared = .repsilon(data))
 
   ## CI
-  ci_method <- NULL
   if (is.numeric(ci)) {
     if (insight::check_if_installed("boot", "for estimating CIs", stop = FALSE)) {
-      out <- cbind(out, .repsilon_ci(data, ci, alternative, iterations))
+      out <- cbind(out, .boot_two_group_es(data, .repsilon, iterations,
+                                           ci, alternative, c(0, 1)))
       ci_method <- list(method = "percentile bootstrap", iterations = iterations)
     } else {
       ci <- NULL
     }
   }
-  if (is.null(ci)) alternative <- NULL
+
+  if (!is.numeric(ci)) {
+    alternative <- NULL
+    ci_method <- NULL
+  }
 
   class(out) <- c("effectsize_table", "see_effectsize_table", class(out))
   attr(out, "ci") <- ci
@@ -122,6 +132,61 @@ rank_epsilon_squared <- function(x,
   attr(out, "alternative") <- alternative
   return(out)
 }
+
+#' @export
+#' @rdname rank_epsilon_squared
+#' @importFrom stats na.omit
+#' @importFrom insight check_if_installed
+rank_eta_squared <- function(x,
+                             groups,
+                             data = NULL,
+                             ci = 0.95,
+                             alternative = "greater",
+                             iterations = 200,
+                             ...) {
+  alternative <- match.arg(alternative, c("greater", "two.sided", "less"))
+
+  if (inherits(x, "htest")) {
+    if (!grepl("Kruskal-Wallis", x$method)) {
+      stop("'x' is not a Kruskal-Wallis-test!", call. = FALSE)
+    }
+    return(effectsize(x, type = "eta", ci = ci, iterations = iterations, alternative = alternative))
+  }
+
+  ## pep data
+  data <- .get_data_multi_group(x, groups, data, ...)
+  data <- stats::na.omit(data)
+
+  out <- data.frame(rank_eta_squared = .reta(data))
+
+  ## CI
+  if (is.numeric(ci)) {
+    if (insight::check_if_installed("boot", "for estimating CIs", stop = FALSE)) {
+      out <- cbind(out, .boot_two_group_es(data, .reta, iterations,
+                                           ci, alternative, c(0, 1)))
+      ci_method <- list(method = "percentile bootstrap", iterations = iterations)
+    } else {
+      ci <- NULL
+    }
+  }
+
+  if (!is.numeric(ci)) {
+    alternative <- NULL
+    ci_method <- NULL
+  }
+
+  class(out) <- c("effectsize_table", "see_effectsize_table", class(out))
+  attr(out, "ci") <- ci
+  attr(out, "ci_method") <- ci_method
+  attr(out, "approximate") <- FALSE
+  attr(out, "alternative") <- alternative
+  return(out)
+}
+
+
+
+
+
 
 #' @rdname rank_epsilon_squared
 #' @export
@@ -156,7 +221,6 @@ kendalls_w <- function(x,
   out <- data.frame(Kendalls_W = W)
 
   ## CI
-  ci_method <- NULL
   if (is.numeric(ci)) {
     if (insight::check_if_installed("boot", "for estimating CIs", stop = FALSE)) {
       out <- cbind(out, .kendalls_w_ci(data, ci, alternative, iterations))
@@ -165,7 +229,11 @@ kendalls_w <- function(x,
       ci <- NULL
     }
   }
-  if (is.null(ci)) alternative <- NULL
+
+  if (!is.numeric(ci)) {
+    alternative <- NULL
+    ci_method <- NULL
+  }
 
   class(out) <- c("effectsize_table", "see_effectsize_table", class(out))
   attr(out, "ci") <- ci
@@ -210,12 +278,24 @@ kendalls_w <- function(x,
 #' @keywords internal
 #' @importFrom stats kruskal.test
 .repsilon <- function(data) {
-  model <- stats::kruskal.test(data$x, data$groups)
+  model <- suppressWarnings(stats::kruskal.test(data$x, data$groups))
 
   H <- unname(model$statistic)
   n <- nrow(data)
 
   E <- H / ((n^2 - 1) / (n + 1))
+}
+
+#' @keywords internal
+#' @importFrom stats kruskal.test
+.reta <- function(data) {
+  model <- suppressWarnings(stats::kruskal.test(data$x, data$groups))
+
+  k <- length(levels(data$groups))
+  n <- nrow(data)
+  E <- model$statistic
+
+  (E - k + 1)/(n - k)
 }
 
 
@@ -243,13 +323,10 @@ kendalls_w <- function(x,
       )
     }
 
-    Tj <- 0
-    for (i in seq_len(m)) {
-      rater <- table(rankings[i, ])
-      ties <- rater[rater > 1]
-      l <- as.numeric(ties)
-      Tj <- Tj + sum(l^3 - l)
-    }
+    Tj <- sum(apply(rankings, 1, function(.r) {
+      TTi <- table(.r)
+      sum(TTi ^ 3 - TTi)
+    }))
 
     W <- (12 * sum(R^2) - 3 * (m^2) * n * ((n + 1)^2)) /
       (m^2 * (n^3 - n) - m * Tj)
@@ -263,26 +340,24 @@ kendalls_w <- function(x,
 
 ## CI ----
 
-
-
 #' @importFrom utils tail
 #' @keywords internal
-.repsilon_ci <- function(data, ci, alternative, iterations) {
+.boot_two_group_es <- function(data, foo_es, iterations,
+                               ci, alternative, lim) {
   stopifnot(length(ci) == 1, ci < 1, ci > 0)
   ci.level <- if (alternative == "two.sided") ci else 2 * ci - 1
 
-
-  boot_r_epsilon <- function(.data, .i) {
-    split(.data$x, .data$groups) <- lapply(split(.data$x, .data$groups),
-                                           sample,
-                                           replace = TRUE
-    )
-    .repsilon(.data)
+  boot_fun <- function(.data, .i) {
+    split(.data$x, .data$groups) <-
+      lapply(split(.data$x, .data$groups),
+             sample,
+             replace = TRUE)
+    foo_es(.data)
   }
 
   R <- boot::boot(
     data = data,
-    statistic = boot_r_epsilon,
+    statistic = boot_fun,
     R = iterations
   )
 
@@ -291,8 +366,8 @@ kendalls_w <- function(x,
 
   data.frame(
     CI = ci,
-    CI_low = if (alternative == "less") 0 else bCI[1],
-    CI_high = if (alternative == "greater") 1 else bCI[2]
+    CI_low = if (alternative == "less") lim[1] else bCI[1],
+    CI_high = if (alternative == "greater") lim[2] else bCI[2]
   )
 }